StickTatorUSBClient/requirements.txt
asyncio==3.4.3
python-dotenv==1.0.1
websockets==15.0
